zebconf: Create an abstract notion of a Zebra device connection
Mac OS X does not provide a character device abstraction for USB
printer devices, and will therefore require the use of libusb in order
to configure a locally attached printer.  Allow for this possibility
by separating out the assumption of the existence of a file handle to
a separate device connection class.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>